% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivo_plano_srv.R
\name{archivo_plano_srv}
\alias{archivo_plano_srv}
\title{Leer archivos planos del server:}
\usage{
archivo_plano_srv(ruta)
}
\arguments{
\item{ruta}{Texto con la ruta específica del archivo a leer (ej.: "aerocomercial/anac/tabla_final.txt")}
}
\description{
Hace una llamada getUrl a la ruta indicada dentro de /srv/DataDNMYE/. Usarlo con read_csv o read_csv2 u otros.
}
\examples{
Devuelve las lineas del texto.
archivo_plano_srv("turismo_internacional/termometro/perfiles_gasto.csv")
Lee las lineas a un tibble.
read_csv2(archivo_plano_srv("turismo_internacional/termometro/perfiles_gasto.csv"))

}
