% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_proyecto.R
\name{crear_proyecto}
\alias{crear_proyecto}
\title{Crear esqueleto de un nuevo proyecto de R}
\usage{
crear_proyecto(ruta = NULL)
}
\arguments{
\item{ruta}{nombre de la carpeta y proyecto que se quiere crear}
}
\value{
directorio base para un nuevo proyecto
}
\description{
Crear esqueleto de un nuevo proyecto de R
}
