% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{\%out\%}
\alias{\%out\%}
\title{Negación de \%in\% :}
\usage{
a \%out\% b
}
\arguments{
\item{a}{vector con conjunto de datos}

\item{b}{vetor con conjunto de datos 2}
}
\description{
elementos no contenidos dentro de un conjunto
}
