% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herramientas.R
\docType{package}
\name{herramientas}
\alias{herramientas}
\title{\code{herramientas} package}
\description{
Paquete de R con utilidades para "domar los datos" de la DNMYE
See the README on
\href{https://github.com/dnme-minturdep/herramientas/blob/main/README.md}{Github}
}
