% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_texto.R
\name{limpiar_texto}
\alias{limpiar_texto}
\title{Limpieza general de texto:}
\usage{
limpiar_texto(texto, enie = T)
}
\arguments{
\item{texto}{variable de clase 'character' a la que se le quiere aplicar la función}

\item{enie}{valor lógico (TRUE o FALSE) indicando si reemplazar "ñ" por "n"}
}
\description{
Aplica tolower(), str_replace_all() para quitar caracteres especiales y remover_tildes().
También aplica opcionalmente str_replace_all() convertir "ñ" en "n" (por default se aplica). Ej.:
}
\examples{
limpiar_texto("TéxTÔ con una Ñ?!")
limpiar_texto("TéxTÔ con una Ñ?!", enie = F)
}
