% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_srv.R
\name{ls_srv}
\alias{ls_srv}
\title{Listar carpetas y archivos del server:}
\usage{
ls_srv(ruta = NULL)
}
\arguments{
\item{ruta}{Texto con la ruta específica del archivo a leer (ej.: "aerocomercial/anac/base_anac_agrupada.rds")}
}
\description{
Devuelve un vector con las ubicaciones en la ruta indicada dentro de /srv/DataDNMYE/.
}
\examples{
ls_srv()
ls_srv("peajes")

}
