% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_srv.R
\name{read_file_srv}
\alias{read_file_srv}
\title{Leer archivos del server:}
\usage{
read_file_srv(ruta, ...)
}
\arguments{
\item{ruta}{Texto con la ruta específica del archivo a leer con su extensión (acepta los siguientes formatos: "rds", "csv", "sav", "txt", ".parquet", "xlsx", "xls", "gpkg", "geojson" y "kml") . Ej.: "aerocomercial/anac/base_anac_agrupada.rds".}

\item{...}{Parámetros para pasarle a la función de lectura subyacente:
\itemize{
\item csv/txt: readr::read_delim
\item rds: base::readRDS
\item sav: readr::read_sav
\item xlsx/xls: readxl::read_excel
\item parquet: arrow::read_parquet
\item gpkg/geojson/kml: sf::read_sf
}}
}
\description{
Hace una conexión a la ruta indicada dentro de /srv/DataDNMYE/ y lee el archivo con una función acorde a la extensión del mismo.
}
