% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds_srv.R
\name{read_rds_srv}
\alias{read_rds_srv}
\title{Leer archivos rds del server:}
\usage{
read_rds_srv(ruta)
}
\arguments{
\item{ruta}{Texto con la ruta específica del archivo a leer (ej.: "aerocomercial/anac/base_anac_agrupada.rds")}
}
\description{
Hace una conexión a la ruta indicada dentro de /srv/DataDNMYE/ y lee el archivo RDS con readRDS().
}
\examples{
read_rds_srv(ruta = "aerocomercial/anac/base_anac_agrupada.rds")

}
