% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav_srv.R
\name{read_sav_srv}
\alias{read_sav_srv}
\title{Leer archivos sav del server:}
\usage{
read_sav_srv(ruta)
}
\arguments{
\item{ruta}{Texto con la ruta específica del archivo a leer (ej.: "evyth/microdatos/evyth_microdatos.sav")}
}
\description{
Hace una conexión a la ruta indicada dentro de /srv/DataDNMYE/ y lee el archivo .sav con haven::read_sav().
}
\examples{
read_sav_srv("evyth/microdatos/evyth_microdatos.sav")

}
