% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remover_tildes.R
\name{remover_tildes}
\alias{remover_tildes}
\title{Limpieza general de tildes:}
\usage{
remover_tildes(texto)
}
\arguments{
\item{texto}{variable de clase 'character' a la que se le quiere aplicar la función}
}
\description{
Elimina tildes las vocales y la letra "y". Ej.: "ÿúòâ" en "yuoa".
}
