% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_proyecto_git.R
\name{sync_proyecto_git}
\alias{sync_proyecto_git}
\title{Sincronizar proyecto con Github}
\usage{
sync_proyecto_git(repo = "d4t4tur", privado = TRUE, protocolo = "ssh")
}
\arguments{
\item{repo}{Por defecto estos se crean en repositorio DEP ('d4t4tur'). Se puede pasar 'repo = "dnme-minturdep"'
como argumento alternativo para vincular con repositorio PROD}

\item{privado}{Por defecto el repositorio será privado. Se puede pasar como argumento
'privado = FALSE' para que sea público'}

\item{protocolo}{Por defecto el repositorio usa protocolo de conexión 'ssh'. Se puede pasar como argumento
'protocolo = "https"' como alternativa}
}
\value{
Vínculo origin/remote para un repositorio .git sincronizado con Organización en Github
}
\description{
Vincula proyecto de R local con repositorio de organización en Github.
}
