% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasa_variacion.R
\name{tasa_variacion}
\alias{tasa_variacion}
\title{Tasa de variación:}
\usage{
tasa_variacion(
  variable,
  periodo = "anual",
  porcentaje = FALSE,
  presentacion = FALSE
)
}
\arguments{
\item{variable}{variable numérica para calcular la variación}

\item{porcentaje}{variación en términos porcentuales (multiplica por 100)}

\item{presentacion}{resultado en formato character con separador de decimales y miles usuales
tasa_varaicion(c(10,10,10,10))
tasa_variacion(c(10,10,10,10),periodo = trimestral)}

\item{período}{indicar cuántos períodos para atrás se calcula. Por defecto es anual (1 período),
puede elegirse "mensual" o "trimestral" o directamente el número (12 para mensual por ej.)}
}
\description{
Calcula la tasa de variación con respecto al período previo según se defina, aplicando dplyr::lag()
}
