% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_tree_srv.R
\name{view_tree_srv}
\alias{view_tree_srv}
\title{Diagrama de carpetas y archivos del server:}
\usage{
view_tree_srv()
}
\description{
Devuelve un gráfico interactivo con los directorios y archivos en /srv/DataDNMYE/. El último nodo contiene la ruta completa de la carpeta o archivo para poder copiar y pegar.
}
