% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetar_provincia.R
\name{etiquetar_provincia}
\alias{etiquetar_provincia}
\title{Etiquetado de variable provincias:}
\usage{
etiquetar_provincia(base, id_col, add_codes = T)
}
\arguments{
\item{base}{Dataframe con una variable que refiere a las provincias}

\item{id_col}{Columna con el nombre de la provincia o su código INDEC}

\item{add_codes}{Indica si se desea agregar la columna con el código de la provincia. Por default es TRUE}
}
\description{
Genera una columna, denominada "provincia_nombre", con el nombre oficial de la provincia y otra columna, denominada "provincia_codigo", con su código numérico INDEC (dos posiciones)
}
