% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numero_en_palabras.R
\name{numeros_en_palabras}
\alias{numeros_en_palabras}
\title{Limpieza de digitos en texto:}
\usage{
numeros_en_palabras(texto, intervalo = 0:100)
}
\arguments{
\item{texto}{variable de clase 'character' a la que se le quiere aplicar la función}

\item{intervalo}{rango entre 0 y 100 con los numeros que se desean reemplazar. Ej. intervalo = 37:50}
}
\description{
Transforma digitos del 0 al 100 en palabras. Ej.: "1" en "uno", "25" en "veinticinco".
}
\examples{
numeros_en_palabras("25 de mayo") #"veinticinco de mayo"
numeros_en_palabras("6 de abril del 98") #"seis de abril del noventa y ocho"

}
