% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_srv.R
\name{read_file_srv}
\alias{read_file_srv}
\title{Leer archivos del server:}
\usage{
read_file_srv(ruta, sheet = NULL)
}
\arguments{
\item{ruta}{Texto con la ruta específica del archivo a leer con su extensión (acepta los siguientes formatos: "rds", "csv", "sav", "txt", ".parquet", "xlsx" y "xls") . Ej.: "aerocomercial/anac/base_anac_agrupada.rds".}

\item{sheet}{En el caso de leer archivos excel, el número o nombre de la pestaña. Por default levanta la primera}
}
\description{
Hace una conexión a la ruta indicada dentro de /srv/DataDNMYE/ y lee el archivo con una función acorde a la extensión del mismo.
}
