% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file_srv.R
\name{write_file_srv}
\alias{write_file_srv}
\title{Leer archivos rds del server:}
\usage{
write_file_srv(x, ruta)
}
\arguments{
\item{x}{Objeto a escribir (ej: data.frame)}

\item{ruta}{Texto con la ruta del archivo, incluyendo nombre y extensión del mismo. Ej: "aerocomercial/anac/base_final.csv")}
}
\description{
Hace una conexión a la ruta indicada dentro de /srv/DataDNMYE/ y escribe un archivo CSV o RDS.
}
